Final touches to error estimation in MCR.

+++ trunk/src/LiBELa/RunEngine.cpp
@@ -986,13 +986,13 @@ void TEMP_SCHEME::mcr_run(){
 
             volume = (EqMC->XSize*EqMC->YSize*EqMC->ZSize);
 
-            sprintf(info, "MCR %7d %10.4f %10.4g %10.4g %10.4g %10.4Lg %10.4g %10.4g %10.4g", i+1, Input->mcr_coefficients[i], bt, EqMC->average_energy, EqMC->energy_standard_deviation, EqMC->MCR_Boltzmann_weighted_average,
-                    log(double(EqMC->MCR_Boltzmann_weighted_average)), (1/log(double(EqMC->MCR_Boltzmann_weighted_average)))*EqMC->MCR_Boltzmann_weighted_stdev ,volume);
+            sprintf(info, "MCR %7d %10.4f %10.4g %10.4g %10.4g %10.4Lg %10.4g %10.4Lg %10.4g", i+1, Input->mcr_coefficients[i], bt, EqMC->average_energy, EqMC->energy_standard_deviation, EqMC->MCR_Boltzmann_weighted_average,
+                    log(double(EqMC->MCR_Boltzmann_weighted_average)), (1/log(double(EqMC->MCR_Boltzmann_weighted_average)))*EqMC->MCR_Boltzmann_weighted_stdev , volume);
             Writer->print_info(info);
             Writer->print_line();
 
             cum_W += (log(double(EqMC->MCR_Boltzmann_weighted_average)));
-            cum_W_err += (1/log(double(EqMC->MCR_Boltzmann_weighted_average)))*EqMC->MCR_Boltzmann_weighted_stdev;
+            cum_W_err += double((1/log(double(EqMC->MCR_Boltzmann_weighted_average)))*EqMC->MCR_Boltzmann_weighted_stdev);
             if (volume > max_vol){
                 max_vol=volume;
             }
@@ -1025,12 +1025,12 @@ void TEMP_SCHEME::mcr_run(){
                 }
                 EqMC->ligand_run(RefLig, LIG, LIG->xyz, Input, bt);
                 lig_volume = (EqMC->XSize*EqMC->YSize*EqMC->ZSize);
-                sprintf(info, "MCR %7d %10.4f %10.4g %10.4g %10.4g %10.4Lg %10.4g %10.4g %10.4g", i+1, Input->mcr_coefficients[i], bt, EqMC->average_energy, EqMC->energy_standard_deviation, EqMC->MCR_Boltzmann_weighted_average,
+                sprintf(info, "MCR %7d %10.4f %10.4g %10.4g %10.4g %10.4Lg %10.4g %10.4Lg %10.4g", i+1, Input->mcr_coefficients[i], bt, EqMC->average_energy, EqMC->energy_standard_deviation, EqMC->MCR_Boltzmann_weighted_average,
                         log(double(EqMC->MCR_Boltzmann_weighted_average)), (1/log(double(EqMC->MCR_Boltzmann_weighted_average)))*EqMC->MCR_Boltzmann_weighted_stdev , lig_volume);
                 Writer->print_info(info);
 
                 cum_W_lig += (log(double(EqMC->MCR_Boltzmann_weighted_average)));
-                cum_W_lig_err += (1/log(double(EqMC->MCR_Boltzmann_weighted_average)))*EqMC->MCR_Boltzmann_weighted_stdev;
+                cum_W_lig_err += double((1/log(double(EqMC->MCR_Boltzmann_weighted_average)))*EqMC->MCR_Boltzmann_weighted_stdev);
                 if (volume > max_vol){
                     lig_max_vol=lig_volume;
                 }
